requirements.txt


aiofiles==0.4.0
aiohttp==3.5.4
asyncio==3.4.3
fastai==1.0.60
torchvision==0.5.0
numpy==1.16.3
starlette==0.12.0
uvicorn==0.7.1
python-multipart==0.0.5



# torchvision==0.2.2
# fastai==1.0.51

# torch==1.4.0
